proto库整理

+++ Source/ThirdPersonMP/Base/Define/DefineVariable.h
@@ -6,8 +6,8 @@ typedef char CHAR;
 typedef unsigned char BYTE;
 typedef unsigned short WORD;
 typedef short SWORD;
-typedef unsigned long long QWORD;
-typedef long long SQWORD;
+typedef uint64_t QWORD;
+typedef int64_t SQWORD;
 typedef float FLOAT;
 typedef double DOUBLE;
